[refactor]: 파일 및 경로 이름 수정 완료 (#53)
